Refactor active detention requests into a separate table
