+++ apps/judicial-system/web/src/routes/Shared/DetentionRequests/ActiveDetentionRequests.tsx
@@ -1,0 +1,279 @@
+import React, { useContext, useMemo, useState } from 'react'
+import cn from 'classnames'
+import { Box, Text, Tag, Icon, Button } from '@island.is/island-ui/core'
+
+import * as styles from './DetentionRequests.treat'
+import {
+  getClassNamesFor,
+  handleClick,
+  mapCaseStateToTagVariant,
+} from './utils'
+import {
+  Case,
+  CaseState,
+  CaseType,
+  UserRole,
+} from '@island.is/judicial-system/types'
+import { insertAt } from '@island.is/judicial-system-web/src/utils/formatters'
+import format from 'date-fns/format'
+import parseISO from 'date-fns/parseISO'
+import { UserContext } from '@island.is/judicial-system-web/src/shared-components/UserProvider/UserProvider'
+import {
+  directionType,
+  SortConfig,
+} from '@island.is/judicial-system-web/src/types'
+import localeIS from 'date-fns/locale/is'
+
+interface Props {
+  cases: Case[]
+  onDeleteCase?: (caseToDelete: Case) => Promise<void>
+  requestToRemoveIndex?: number
+  handleDeleteButtonClick?: (index: number) => void
+}
+
+const ActiveDetentionRequests: React.FC<Props> = (props) => {
+  const {
+    cases,
+    onDeleteCase,
+    requestToRemoveIndex,
+    handleDeleteButtonClick,
+  } = props
+
+  const { user } = useContext(UserContext)
+  const isProsecutor = user?.role === UserRole.PROSECUTOR
+  const isJudge = user?.role === UserRole.JUDGE
+
+  const [sortConfig, setSortConfig] = useState<SortConfig>()
+
+  useMemo(() => {
+    const sortedCases = cases || []
+
+    if (sortConfig) {
+      sortedCases.sort((a: Case, b: Case) => {
+        // Credit: https://stackoverflow.com/a/51169
+        return sortConfig.direction === 'ascending'
+          ? ('' + a[sortConfig.key]).localeCompare(
+              b[sortConfig.key]?.toString() || '',
+            )
+          : ('' + b[sortConfig.key]).localeCompare(
+              a[sortConfig.key]?.toString() || '',
+            )
+      })
+    }
+    return sortedCases
+  }, [cases, sortConfig])
+
+  const requestSort = (key: keyof Case) => {
+    let d: directionType = 'ascending'
+
+    if (
+      sortConfig &&
+      sortConfig.key === key &&
+      sortConfig.direction === 'ascending'
+    ) {
+      d = 'descending'
+    }
+    setSortConfig({ key, direction: d })
+  }
+
+  return (
+    <table
+      className={styles.activeRequestsTable}
+      data-testid="detention-requests-table"
+      aria-describedby="activeRequestsTableCaption"
+    >
+      <thead className={styles.thead}>
+        <tr>
+          <th className={styles.th}>
+            <Text as="span" fontWeight="regular">
+              Málsnr.
+            </Text>
+          </th>
+          <th className={cn(styles.th, styles.largeColumn)}>
+            <Box
+              component="button"
+              display="flex"
+              alignItems="center"
+              className={styles.thButton}
+              onClick={() => requestSort('accusedName')}
+              data-testid="accusedNameSortButton"
+            >
+              <Text fontWeight="regular">Sakborningur</Text>
+              <Box
+                className={cn(styles.sortIcon, {
+                  [styles.sortAccusedNameAsc]:
+                    getClassNamesFor('accusedName', sortConfig) === 'ascending',
+                  [styles.sortAccusedNameDes]:
+                    getClassNamesFor('accusedName', sortConfig) ===
+                    'descending',
+                })}
+                marginLeft={1}
+                component="span"
+                display="flex"
+                alignItems="center"
+              >
+                <Icon icon="caretDown" size="small" />
+              </Box>
+            </Box>
+          </th>
+          <th className={styles.th}>
+            <Text as="span" fontWeight="regular">
+              Tegund
+            </Text>
+          </th>
+          <th className={styles.th}>
+            <Text as="span" fontWeight="regular">
+              Staða
+            </Text>
+          </th>
+          <th className={styles.th}>
+            <Box
+              component="button"
+              display="flex"
+              alignItems="center"
+              className={styles.thButton}
+              onClick={() => requestSort('created')}
+            >
+              <Text fontWeight="regular">Krafa stofnuð</Text>
+              <Box
+                className={cn(styles.sortIcon, {
+                  [styles.sortCreatedAsc]:
+                    getClassNamesFor('created', sortConfig) === 'ascending',
+                  [styles.sortCreatedDes]:
+                    getClassNamesFor('created', sortConfig) === 'descending',
+                })}
+                marginLeft={1}
+                component="span"
+                display="flex"
+                alignItems="center"
+              >
+                <Icon icon="caretUp" size="small" />
+              </Box>
+            </Box>
+          </th>
+          <th></th>
+        </tr>
+      </thead>
+      <tbody>
+        {cases.map((c, i) => (
+          <tr
+            key={i}
+            className={cn(
+              styles.tableRowContainer,
+              requestToRemoveIndex === i && 'isDeleting',
+            )}
+            data-testid="detention-requests-table-row"
+            role="button"
+            aria-label="Opna kröfu"
+            onClick={() => {
+              handleClick(c.state, c.id, user?.role, c.isCourtDateInThePast)
+            }}
+          >
+            <td className={styles.td}>
+              <Text as="span">{c.policeCaseNumber || '-'}</Text>
+            </td>
+            <td className={cn(styles.td, styles.largeColumn)}>
+              <Text>
+                <Box component="span" className={styles.accusedName}>
+                  {c.accusedName || '-'}
+                </Box>
+              </Text>
+              <Text>
+                {c.accusedNationalId && (
+                  <Text as="span" variant="small" color="dark400">
+                    {`kt: ${
+                      insertAt(c.accusedNationalId.replace('-', ''), '-', 6) ||
+                      '-'
+                    }`}
+                  </Text>
+                )}
+              </Text>
+            </td>
+            <td className={styles.td}>
+              <Box component="span" display="flex" flexDirection="column">
+                <Text as="span">
+                  {c.type === CaseType.CUSTODY ? 'Gæsluvarðhald' : 'Farbann'}
+                </Text>
+                {c.parentCase && (
+                  <Text as="span" variant="small" color="dark400">
+                    Framlenging
+                  </Text>
+                )}
+              </Box>
+            </td>
+            <td className={styles.td} data-testid="tdTag">
+              <Tag
+                variant={
+                  mapCaseStateToTagVariant(
+                    c.state,
+                    isJudge,
+                    c.isCustodyEndDateInThePast,
+                  ).color
+                }
+                outlined
+                disabled
+              >
+                {
+                  mapCaseStateToTagVariant(
+                    c.state,
+                    isJudge,
+                    c.isCustodyEndDateInThePast,
+                  ).text
+                }
+              </Tag>
+            </td>
+            <td className={styles.td}>
+              <Text as="span">
+                {format(parseISO(c.created), 'd.M.y', {
+                  locale: localeIS,
+                })}
+              </Text>
+            </td>
+            <td className={cn(styles.td, 'secondLast')}>
+              {isProsecutor &&
+                (c.state === CaseState.NEW ||
+                  c.state === CaseState.DRAFT ||
+                  c.state === CaseState.SUBMITTED ||
+                  c.state === CaseState.RECEIVED) && (
+                  <Box
+                    data-testid="deleteCase"
+                    component="button"
+                    aria-label="Viltu afturkalla kröfu?"
+                    className={styles.deleteButton}
+                    onClick={(evt) => {
+                      evt.stopPropagation()
+                      handleDeleteButtonClick && handleDeleteButtonClick(i)
+                    }}
+                  >
+                    <Icon icon="close" color="blue400" />
+                  </Box>
+                )}
+            </td>
+            <td
+              className={cn(
+                styles.deleteButtonContainer,
+                styles.td,
+                requestToRemoveIndex === i && 'open',
+              )}
+            >
+              <Button
+                colorScheme="destructive"
+                size="small"
+                onClick={(evt) => {
+                  evt.stopPropagation()
+                  onDeleteCase && onDeleteCase(cases[i])
+                }}
+              >
+                <Box as="span" className={styles.deleteButtonText}>
+                  Afturkalla
+                </Box>
+              </Button>
+            </td>
+          </tr>
+        ))}
+      </tbody>
+    </table>
+  )
+}
+
+export default ActiveDetentionRequests
