+++ apps/judicial-system/web/src/routes/Shared/DetentionRequests/DetentionRequests.tsx
@@ -1,13 +1,5 @@
 import React, { useEffect, useState, useContext, useMemo } from 'react'
-import cn from 'classnames'
-import {
-  AlertMessage,
-  Box,
-  Text,
-  Tag,
-  Icon,
-  Button,
-} from '@island.is/island-ui/core'
+import { AlertMessage, Box, Text, Tag } from '@island.is/island-ui/core'
 import {
   DropdownMenu,
   Loading,
@@ -23,10 +15,7 @@ import {
 } from '@island.is/judicial-system/types'
 import { UserRole } from '@island.is/judicial-system/types'
 import * as Constants from '@island.is/judicial-system-web/src/utils/constants'
-import {
-  insertAt,
-  parseTransition,
-} from '@island.is/judicial-system-web/src/utils/formatters'
+import { parseTransition } from '@island.is/judicial-system-web/src/utils/formatters'
 import { useMutation, useQuery } from '@apollo/client'
 import { UserContext } from '@island.is/judicial-system-web/src/shared-components/UserProvider/UserProvider'
 import {
@@ -37,23 +26,13 @@ import { CasesQuery } from '@island.is/judicial-system-web/src/utils/mutations'
 import * as styles from './DetentionRequests.treat'
 import { formatDate } from '@island.is/judicial-system/formatters'
 import parseISO from 'date-fns/parseISO'
-import {
-  getClassNamesFor,
-  handleClick,
-  mapCaseStateToTagVariant,
-} from './utils'
-import format from 'date-fns/format'
-import localeIS from 'date-fns/locale/is'
-import {
-  directionType,
-  SortConfig,
-} from '@island.is/judicial-system-web/src/types'
+import { handleClick, mapCaseStateToTagVariant } from './utils'
+import ActiveDetentionRequests from './ActiveDetentionRequests'
 
 // Credit for sorting solution: https://www.smashingmagazine.com/2020/03/sortable-tables-react/
 export const DetentionRequests: React.FC = () => {
   const [activeCases, setActiveCases] = useState<Case[]>()
   const [pastCases, setPastCases] = useState<Case[]>()
-  const [sortConfig, setSortConfig] = useState<SortConfig>()
 
   // The index of requset that's about to be removed
   const [requestToRemoveIndex, setRequestToRemoveIndex] = useState<number>()
@@ -131,24 +110,6 @@ export const DetentionRequests: React.FC = () => {
     isRegistrar,
     resCases,
   ])
-
-  useMemo(() => {
-    const sortedCases = activeCases || []
-
-    if (sortConfig) {
-      sortedCases.sort((a: Case, b: Case) => {
-        // Credit: https://stackoverflow.com/a/51169
-        return sortConfig.direction === 'ascending'
-          ? ('' + a[sortConfig.key]).localeCompare(
-              b[sortConfig.key]?.toString() || '',
-            )
-          : ('' + b[sortConfig.key]).localeCompare(
-              a[sortConfig.key]?.toString() || '',
-            )
-      })
-    }
-    return sortedCases
-  }, [activeCases, sortConfig])
 
   const pastRequestsColumns = useMemo(
     () => [
@@ -249,8 +210,11 @@ export const DetentionRequests: React.FC = () => {
   )
 
   const pastRequestsData = useMemo(() => pastCases, [pastCases])
-
   const sortableColumnIds = ['courtCaseNumber', 'accusedName', 'type']
+
+  const handleDeleteButtonClick = (index: number) => {
+    setRequestToRemoveIndex(index === requestToRemoveIndex ? undefined : index)
+  }
 
   const deleteCase = async (caseToDelete: Case) => {
     if (
@@ -295,19 +259,6 @@ export const DetentionRequests: React.FC = () => {
     }
   }
 
-  const requestSort = (key: keyof Case) => {
-    let d: directionType = 'ascending'
-
-    if (
-      sortConfig &&
-      sortConfig.key === key &&
-      sortConfig.direction === 'ascending'
-    ) {
-      d = 'descending'
-    }
-    setSortConfig({ key, direction: d })
-  }
-
   return (
     <div className={styles.detentionRequestsContainer}>
       {user && (
@@ -345,223 +296,12 @@ export const DetentionRequests: React.FC = () => {
             </Text>
           </Box>
           {activeCases && activeCases.length > 0 ? (
-            <table
-              className={styles.activeRequestsTable}
-              data-testid="detention-requests-table"
-              aria-describedby="activeRequestsTableCaption"
-            >
-              <thead className={styles.thead}>
-                <tr>
-                  <th className={styles.th}>
-                    <Text as="span" fontWeight="regular">
-                      Málsnr.
-                    </Text>
-                  </th>
-                  <th className={cn(styles.th, styles.largeColumn)}>
-                    <Box
-                      component="button"
-                      display="flex"
-                      alignItems="center"
-                      className={styles.thButton}
-                      onClick={() => requestSort('accusedName')}
-                      data-testid="accusedNameSortButton"
-                    >
-                      <Text fontWeight="regular">Sakborningur</Text>
-                      <Box
-                        className={cn(styles.sortIcon, {
-                          [styles.sortAccusedNameAsc]:
-                            getClassNamesFor('accusedName', sortConfig) ===
-                            'ascending',
-                          [styles.sortAccusedNameDes]:
-                            getClassNamesFor('accusedName', sortConfig) ===
-                            'descending',
-                        })}
-                        marginLeft={1}
-                        component="span"
-                        display="flex"
-                        alignItems="center"
-                      >
-                        <Icon icon="caretDown" size="small" />
-                      </Box>
-                    </Box>
-                  </th>
-                  <th className={styles.th}>
-                    <Text as="span" fontWeight="regular">
-                      Tegund
-                    </Text>
-                  </th>
-                  <th className={styles.th}>
-                    <Text as="span" fontWeight="regular">
-                      Staða
-                    </Text>
-                  </th>
-                  <th className={styles.th}>
-                    <Box
-                      component="button"
-                      display="flex"
-                      alignItems="center"
-                      className={styles.thButton}
-                      onClick={() => requestSort('created')}
-                    >
-                      <Text fontWeight="regular">Krafa stofnuð</Text>
-                      <Box
-                        className={cn(styles.sortIcon, {
-                          [styles.sortCreatedAsc]:
-                            getClassNamesFor('created', sortConfig) ===
-                            'ascending',
-                          [styles.sortCreatedDes]:
-                            getClassNamesFor('created', sortConfig) ===
-                            'descending',
-                        })}
-                        marginLeft={1}
-                        component="span"
-                        display="flex"
-                        alignItems="center"
-                      >
-                        <Icon icon="caretUp" size="small" />
-                      </Box>
-                    </Box>
-                  </th>
-                  <th></th>
-                </tr>
-              </thead>
-              <tbody>
-                {activeCases.map((c, i) => (
-                  <tr
-                    key={i}
-                    className={cn(
-                      styles.tableRowContainer,
-                      requestToRemoveIndex === i && 'isDeleting',
-                    )}
-                    data-testid="detention-requests-table-row"
-                    role="button"
-                    aria-label="Opna kröfu"
-                    onClick={() => {
-                      handleClick &&
-                        handleClick(
-                          c.state,
-                          c.id,
-                          user?.role,
-                          c.isCourtDateInThePast,
-                        )
-                    }}
-                  >
-                    <td className={styles.td}>
-                      <Text as="span">{c.policeCaseNumber || '-'}</Text>
-                    </td>
-                    <td className={cn(styles.td, styles.largeColumn)}>
-                      <Text>
-                        <Box component="span" className={styles.accusedName}>
-                          {c.accusedName || '-'}
-                        </Box>
-                      </Text>
-                      <Text>
-                        {c.accusedNationalId && (
-                          <Text as="span" variant="small" color="dark400">
-                            {`kt: ${
-                              insertAt(
-                                c.accusedNationalId.replace('-', ''),
-                                '-',
-                                6,
-                              ) || '-'
-                            }`}
-                          </Text>
-                        )}
-                      </Text>
-                    </td>
-                    <td className={styles.td}>
-                      <Box
-                        component="span"
-                        display="flex"
-                        flexDirection="column"
-                      >
-                        <Text as="span">
-                          {c.type === CaseType.CUSTODY
-                            ? 'Gæsluvarðhald'
-                            : 'Farbann'}
-                        </Text>
-                        {c.parentCase && (
-                          <Text as="span" variant="small" color="dark400">
-                            Framlenging
-                          </Text>
-                        )}
-                      </Box>
-                    </td>
-                    <td className={styles.td} data-testid="tdTag">
-                      <Tag
-                        variant={
-                          mapCaseStateToTagVariant(
-                            c.state,
-                            isJudge,
-                            c.isCustodyEndDateInThePast,
-                          ).color
-                        }
-                        outlined
-                        disabled
-                      >
-                        {
-                          mapCaseStateToTagVariant(
-                            c.state,
-                            isJudge,
-                            c.isCustodyEndDateInThePast,
-                          ).text
-                        }
-                      </Tag>
-                    </td>
-                    <td className={styles.td}>
-                      <Text as="span">
-                        {format(parseISO(c.created), 'd.M.y', {
-                          locale: localeIS,
-                        })}
-                      </Text>
-                    </td>
-                    <td className={cn(styles.td, 'secondLast')}>
-                      {isProsecutor &&
-                        (c.state === CaseState.NEW ||
-                          c.state === CaseState.DRAFT ||
-                          c.state === CaseState.SUBMITTED ||
-                          c.state === CaseState.RECEIVED) && (
-                          <Box
-                            data-testid="deleteCase"
-                            component="button"
-                            aria-label="Viltu afturkalla kröfu?"
-                            className={styles.deleteButton}
-                            onClick={(evt) => {
-                              evt.stopPropagation()
-                              setRequestToRemoveIndex &&
-                                setRequestToRemoveIndex(
-                                  requestToRemoveIndex === i ? undefined : i,
-                                )
-                            }}
-                          >
-                            <Icon icon="close" color="blue400" />
-                          </Box>
-                        )}
-                    </td>
-                    <td
-                      className={cn(
-                        styles.deleteButtonContainer,
-                        styles.td,
-                        requestToRemoveIndex === i && 'open',
-                      )}
-                    >
-                      <Button
-                        colorScheme="destructive"
-                        size="small"
-                        onClick={(evt) => {
-                          evt.stopPropagation()
-                          deleteCase(activeCases[i])
-                        }}
-                      >
-                        <Box as="span" className={styles.deleteButtonText}>
-                          Afturkalla
-                        </Box>
-                      </Button>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
+            <ActiveDetentionRequests
+              cases={activeCases}
+              onDeleteCase={deleteCase}
+              requestToRemoveIndex={requestToRemoveIndex}
+              handleDeleteButtonClick={handleDeleteButtonClick}
+            />
           ) : (
             <div className={styles.activeRequestsTableInfo}>
               <AlertMessage
